Rename package to mtgjson4 to avoid clash with https://pypi.org/project/mtgjson/
