+++ mtgjson4/shared_info.py
